+++ public/move_drawer_event.js
@@ -1,0 +1,5 @@
+export class MoveDrawerEvent extends CustomEvent {
+    constructor(command) {
+        super('Drawer:move', {detail: {command: command}});
+    }
+}